cybersecurity/requirements.txt
# cybersecurity — AI-Native Security Operations Agent
# Install: pip install -r requirements.txt

# ── LLM & Agent Framework ─────────────────────────────────────────────────────
# OpenAI SDK — primary LLM client (tool calling API used for agent loop)
openai>=1.0.0
# LangChain — used for @tool decorator and tool utilities only
langchain>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# ── Vector DB (RAG) ───────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# ── Data validation ───────────────────────────────────────────────────────────
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Eval Dashboard ────────────────────────────────────────────────────────────
streamlit>=1.30.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.0.0
